refactor: move ast::generic to prql_ast
