+++ crates/prql_ast/src/expr.rs
@@ -1,3 +1,4 @@
+pub mod generic;
 mod ident;
 mod literal;
 mod ops;
