+++ prql-compiler/src/ast/mod.rs
@@ -1,8 +1,8 @@
 //! Abstract Syntax Tree
 //!
-pub mod generic;
 pub mod pl;
 pub mod rq;
 mod span;
 
+pub use prql_ast::expr::generic;
 pub use span::Span;
